Externalize inventoryDiffs query
